Use OAuth2 to authorize endpoints
