+++ amarillo/routers/agencyconf.py
@@ -5,6 +5,7 @@ from fastapi import APIRouter, HTTPException, status, Header, Depends
 
 from amarillo.models.AgencyConf import AgencyConf
 from amarillo.services.agencyconf import AgencyConfService
+from amarillo.services.oauth2 import get_current_agency, verify_admin
 from amarillo.services.config import config
 from amarillo.utils.container import container
 
@@ -19,24 +20,6 @@ router = APIRouter(
 # TODO make this an explicit config option
 include_in_schema = config.env != 'PROD'
 
-
-# noinspection PyPep8Naming
-# X_API_Key is upper case for OpenAPI
-async def verify_admin_api_key(X_API_Key: str = Header(...)):
-    if X_API_Key != config.admin_token:
-        message="X-API-Key header invalid"
-        logger.error(message)
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail=message)
-
-    return "admin"
-
-
-# noinspection PyPep8Naming
-# X_API_Key is upper case for OpenAPI
-async def verify_api_key(X_API_Key: str = Header(...)):
-    agency_conf_service: AgencyConfService = container['agencyconf']
-
-    return agency_conf_service.check_api_key(X_API_Key)
 
 # TODO Return code 403 Unauthoized (in response_status_codes as well...)
 async def verify_permission_for_same_agency_or_admin(agency_id_in_path_or_body, agency_id_from_api_key):
@@ -61,7 +44,7 @@ async def verify_permission_for_same_agency_or_admin(agency_id_in_path_or_body, 
             response_model=List[str],
             description="Returns the agency_ids but not the details.",
             status_code=status.HTTP_200_OK)
-async def get_agency_ids(admin_api_key: str = Depends(verify_api_key)) -> [str]:
+async def get_agency_ids(admin_api_key: str = Depends(get_current_agency)) -> [str]:
     return container['agencyconf'].get_agency_ids()
 
 
@@ -69,7 +52,7 @@ async def get_agency_ids(admin_api_key: str = Depends(verify_api_key)) -> [str]:
              include_in_schema=include_in_schema,
              operation_id="postNewAgencyConf",
              summary="Post a new AgencyConf")
-async def post_agency_conf(agency_conf: AgencyConf, admin_api_key: str = Depends(verify_admin_api_key)):
+async def post_agency_conf(agency_conf: AgencyConf, admin_api_key: str = Depends(verify_admin)):
     agency_conf_service: AgencyConfService = container['agencyconf']
     agency_conf_service.add(agency_conf)
 
@@ -81,7 +64,7 @@ async def post_agency_conf(agency_conf: AgencyConf, admin_api_key: str = Depends
                summary="Delete configuration of an agency. Returns true if the token for the agency existed, "
                        "false if it didn't exist."
                )
-async def delete_agency_conf(agency_id: str, requesting_agency_id: str = Depends(verify_api_key)):
+async def delete_agency_conf(agency_id: str, requesting_agency_id: str = Depends(get_current_agency)):
     agency_may_delete_own = requesting_agency_id == agency_id
     admin_may_delete_everything = requesting_agency_id == "admin"
     is_permitted = agency_may_delete_own or admin_may_delete_everything
